add TypesWithMultipliersSection

+++ app/src/main/java/com/arix/pokedex/extensions/NumberExt.kt
@@ -10,7 +10,7 @@ private fun gcd(a: Long, b: Long): Long {
     return if (b == 0L) a else gcd(b, a % b)
 }
 
-fun Float.toFraction(): Pair<Long, Long> {
+fun Double.toFraction(): Pair<Long, Long> {
     val stringValue = this.toString()
     val decimalPlaces = stringValue.substringAfter('.', "").length
